Define a User class that satisfies both Socialite and Laravel Auth.
Laravel's Auth and Socialite have different requirements for the user
object, and the normal usage pattern seems to be to auto create
database users from the Socialite login.  But this should not be
necessary if Oauth2 logins are the only supported login, it is more
to support OAuth2 login as an option (like Phorge itself does).

That pattern is more for public apps that support social media logon,
rather than an enterprise app supporting Oauth2 for single-sign-on.