Define a new auth provider instead of overriding "users".
Previously we overrode the "users" provider (which may have relied on
it being commented out in the config/app.php to work).

In Laravel 11, default config is now internal to Laravel itself, and
this overriding does not work. So instead define a new provider called
"phorge" and give instructions on how to reconfigure the guards to use
it.